refactor: remove unused Cards test file and update blog page tests
