+++ src/app/[locale]/components/Header/components/Navigator/navigator.data.ts
@@ -19,11 +19,6 @@ const items = [
     name_en: 'Career',
     href: Paths.CAREER,
   },
-  {
-    name: 'Contanto',
-    name_en: 'Contact',
-    href: Paths.CONTACT,
-  },
 ]
 
 export default items
