--- src/app/[locale]/contact/page.tsx
@@ -1,9 +1,0 @@
-const Contact = async () => {
-  return (
-    <section>
-      <h1 className='text-title-xgiant text-white'>[IN PROGRESS]</h1>
-    </section>
-  )
-}
-
-export default Contact
